Improve App User API

+++ src/main/java/com/mhealth/userservice/service/UserServiceImpl.java
@@ -1,35 +1,58 @@
 package com.mhealth.userservice.service;
+
+import com.mhealth.userservice.dto.AppUserDTO;
+import com.mhealth.userservice.dto.UpdatePasswordDTO;
 import com.mhealth.userservice.entity.AppUser;
 import com.mhealth.userservice.repository.AppUserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserServiceImpl implements UserService {
 
-    @Autowired
-    private AppUserRepository userRepository;
+    private final AppUserRepository userRepository;
+
+    public UserServiceImpl(AppUserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
 
     @Override
-    public AppUser createUser(AppUser user) {
+    public AppUser createUser(AppUserDTO userDTO) {
+        AppUser user = new AppUser();
+        BeanUtils.copyProperties(userDTO, user);
         return userRepository.save(user);
     }
 
     @Override
     public AppUser getUserById(Long userId) {
-        return userRepository.findById(userId)
-                .orElse(null);
+        return userRepository.findById(userId).orElse(null);
     }
+
     @Override
-    public void saveUser(AppUser user) {
-        userRepository.save(user);
+    public boolean updateUserPassword(Long userId, UpdatePasswordDTO passwordDTO) {
+        Optional<AppUser> optionalUser = userRepository.findById(userId);
+        if (optionalUser.isPresent()) {
+            AppUser user = optionalUser.get();
+            user.setPassword(passwordDTO.getPassword());
+            userRepository.save(user);
+            return true;
+        }
+        return false;
     }
+
     @Override
-    public void deleteUser(Long userId) {
-        userRepository.deleteById(userId);
+    public boolean deleteUser(Long userId) {
+        Optional<AppUser> optionalUser = userRepository.findById(userId);
+        if (optionalUser.isPresent()) {
+            userRepository.deleteById(userId);
+            return true;
+        }
+        return false;
     }
+
     @Override
     public List<AppUser> getAllUsers() {
         return userRepository.findAll();
